Remove legacy describeConformance
