+++ packages/material-ui-lab/src/StaticTimePicker/StaticTimePicker.test.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import TextField from '@material-ui/core/TextField';
-import { describeConformance } from 'test/utils';
+import { describeConformanceV5 } from 'test/utils';
 import StaticTimePicker from './StaticTimePicker';
 import { wrapPickerMount } from '../internal/pickers/test-utils';
 
 describe('<StaticTimePicker />', () => {
-  describeConformance(
+  describeConformanceV5(
     <StaticTimePicker
       onChange={() => {}}
       renderInput={(props) => <TextField {...props} />}
@@ -17,6 +17,8 @@ describe('<StaticTimePicker />', () => {
       refInstanceof: undefined,
       skip: [
         'componentProp',
+        'componentsProp',
+        'themeDefaultProps',
         'mergeClassName',
         'propsSpread',
         // TODO: `ref` is typed but has no effect
