+++ packages/material-ui/src/Grow/Grow.test.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { expect } from 'chai';
 import { spy, useFakeTimers } from 'sinon';
-import { act, createClientRender, describeConformance } from 'test/utils';
+import { act, createClientRender, describeConformanceV5 } from 'test/utils';
 import { createTheme, ThemeProvider } from '@material-ui/core/styles';
 import { Transition } from 'react-transition-group';
 import Grow from '@material-ui/core/Grow';
@@ -15,7 +15,7 @@ describe('<Grow />', () => {
     children: <div />,
   };
 
-  describeConformance(
+  describeConformanceV5(
     <Grow in>
       <div />
     </Grow>,
@@ -25,6 +25,8 @@ describe('<Grow />', () => {
       refInstanceof: window.HTMLDivElement,
       skip: [
         'componentProp',
+        'componentsProp',
+        'themeDefaultProps',
         // react-transition-group issue
         'reactTestRenderer',
       ],
