+++ test/utils/index.js
@@ -1,5 +1,4 @@
 export * from './components';
-export { default as describeConformance } from './describeConformance';
 export { default as describeConformanceV5 } from './describeConformanceV5';
 export { default as describeConformanceUnstyled } from './describeConformanceUnstyled';
 export * from './createClientRender';
